# Optimize coordinate transformations