In https://github.com/mapbox/mapnik-vector-tile/commit/9dd27291ee7d84507a1b466d3d45a7aa84185791#diff-2aabba777685cca077a1c8766e95e711R39 I optimized the `vector_tile_strategy` significantly to avoid the overhead of the function call to `prj_trans.backward` even when it is a no-op at https://github.com/mapnik/mapnik/blob/b285bde44bb009d5c619d393bb4902cf96bd98d1/src/proj_transform.cpp#L187-L188. I think this was a lot faster because of the excessive amount of non-inlinable function overloads in https://github.com/mapnik/mapnik/blob/b285bde44bb009d5c619d393bb4902cf96bd98d1/include/mapnik/proj_transform.hpp#L47-L58.

This commit optimized further by avoiding the `prj_trans.equal()` call per feature. This speeds up the benchmark noticeably:

```
2249.26ms (cpu 2227.56ms)   | boost::geometry::transform
1935.95ms (cpu 1807.54ms)   | transform_visitor with reserve with proj no-op
1775.08ms (cpu 1699.88ms)   | transform_visitor with reserve with no proj function call overhead
```